+++ src/boot.sh
@@ -38,11 +38,19 @@ trap "rm -rf $WORKING_DIR" EXIT
 
 tput lines >/dev/null 2>&1 && export TERMINAL_LINES=$(tput lines);
 
-if [ "$1" = "-v1" ]; then
+if [ "$#" = "0" ]; then
+    perl $TOOL_DIR/main-v2.pl "$@"
+elif [ "$1" = "-v1" ]; then
     shift
     perl $TOOL_DIR/main-v1.pl "$@"
 elif [ "$1" = "-v2" ]; then
     shift
+    perl $TOOL_DIR/main-v2.pl "$@"
+elif [ "$1" = "--help" ]; then
+    perl $TOOL_DIR/main-v2.pl "$@"
+elif [ "$1" = "help" ]; then
+    perl $TOOL_DIR/main-v2.pl "$@"
+elif [ "$1" = "--version" ]; then
     perl $TOOL_DIR/main-v2.pl "$@"
 else
     perl $TOOL_DIR/main-v1.pl "$@"