+++ test/test.sh
@@ -13,6 +13,7 @@ bash ./test/case-wordsflags.sh
 bash ./test/case-addmap.sh
 bash ./test/case-join.sh
 bash ./test/case-datetime.sh
+bash ./test/case-number.sh
 
 diff -ru ./test/expected ./var/test && echo OK